main/xiaozhi-server/requirements.txt
# -*- coding:utf-8 -*-
#--------- 本项目推荐环境是python3.10，以下暂时不推荐升级的依赖
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
#--------- cozepy0.20.0要求websockets<15.0.0，以下暂时不推荐升级的依赖
websockets==14.2
#--------- 以下是可升级的依赖
pyyml==0.0.2
silero_vad==6.1.0
opuslib_next==1.1.5
pydub==0.25.1
funasr==1.2.7
openai==2.8.1
google-generativeai==0.8.5
edge_tts==7.2.3
httpx==0.28.1
aiohttp==3.13.2
aiohttp_cors==0.8.1
ormsgpack==1.12.0
ruamel.yaml==0.18.16
loguru==0.7.3
requests==2.32.5
cozepy==0.20.0
mem0ai==1.0.0
bs4==0.0.2
modelscope==1.23.2
sherpa_onnx==1.12.17
mcp==1.20.0
cnlunar==0.2.0
PySocks==1.7.1
dashscope==1.25.2
baidu-aip==4.16.13
chardet==5.2.0
aioconsole==0.8.2
markitdown==0.1.3
mcp-proxy==0.10.0
PyJWT==2.10.1
psutil==7.0.0
portalocker==3.2.0
Jinja2==3.1.6
vosk==0.3.45
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4